requirements_gpu.txt
tensorflow-gpu
Keras
sklearn
pillow
wandb
imageio
opencv-python
matplotlib
pandas
tqdm
ipywidgets==7.5.1
